Function zng_view_prebuilt::view_process_main

pub fn view_process_main()

Calls the prebuilt view_process_main.

Note that this only needs to be called if the view-process is not built on the same executable, if it is you only need to call zng_env::init! at the beginning of the executable main.

You can also disable start on init by setting the ZNG_VIEW_NO_INIT_START environment variable. In this case you must manually call this function.